+++ weatherly/responses.py
@@ -28,7 +28,8 @@ from typing import (
     Union,
     Optional,
     List, 
-    Iterator,   
+    Iterator,
+    Literal,
 )
 
 __all__ = (
@@ -40,6 +41,8 @@ __all__ = (
     "ForecastHour",
     "AlertData",
     "FutureData",
+    "AstronomicalData",
+    "IPData",
 )
 
 
@@ -663,8 +666,8 @@ class AstronomicalData(APIResponse):
         HTTP status of the response. 200 is OK, and is the most common status.
     code: Optional[:class:`int`]
         Response code. In some cases this can be ``None``
-    location: :class:`LocationData`
-        Location of the requested data
+    location: Optional[:class:`LocationData`]
+        Location of the requested data. Is not ``None`` only if this object is returned from ``astronomy.json`` endpoint.
     sunrise: :class:`str`
         Sunrise local time
     sunset: :class:`str`
@@ -694,7 +697,17 @@ class AstronomicalData(APIResponse):
     ) -> None:
         super().__init__(raw, status, code)
         
-        self.location: LocationData = LocationData(raw["location"], status, code)
+        # there are two types of "raw" data we need to convert
+        # 1. from astronomy.json 
+        # 2. from other types of endpoints, e.g. forecast data
+        # we set location to None to avoid errors
+        self.location: Optional[LocationData] = None
+                
+        if raw.get("astronomy") is not None:
+            # first type
+            self.location = LocationData(raw["location"], status, code)
+            raw = raw["astronomy"]["astro"]
+        
         self.sunrise: str = raw["sunrise"]
         self.sunset: str = raw["sunset"]
         self.moonrise: str = raw["moonrise"]
@@ -702,3 +715,69 @@ class AstronomicalData(APIResponse):
         self.moon_phase: str = raw["moon_phase"]
         self.moon_illumination: int = int(raw["moon_illumination"])
         
+class IPData(APIResponse):
+    """
+    Represents IP Data retuned from IP Lookup API
+    
+    Attributes
+    ------------
+    raw: Dict[:class:`str`, Any]
+        Raw response in a JSON-like format (converted to a python dictionary)
+    status: :class:`int`
+        HTTP status of the response. 200 is OK, and is the most common status.
+    code: Optional[:class:`int`]
+        Response code. In some cases this can be ``None``
+    ip: :class:`str`
+        IP address
+    type: Literal["ipv4", "ipv6"]
+        Type of the IP address
+    continent_code: :class:`str`
+        Continent code
+    continent_name: :class:`str`
+        Continent name
+    country_code: :class:`str`
+        Country code
+    country_name: :class:`str`
+        Name of country
+    is_eu: :class:`bool`
+        Indicating if IP address is in Europe
+    geoname_id: :class:`int`
+        Geoname ID
+    city: :class:`str`
+        City name
+    region: :class:`str`
+        Region name
+    latitude: :class:`float`
+        Latitude in decimal degree
+    longitude: :class:`float`
+        Longitude in decimal degree
+    tz_id: :class:`str`
+        Time zone
+    localtime_epoch: :class:`int`
+        Local time epoch
+    localtime_formatted: :class:`str`
+        Formatted local time string (e.g. 2023-04-30 17:54)
+    """
+    def __init__(
+        self,
+        raw: Dict[str, Any],
+        status: int,
+        code: Optional[int]
+    ) -> None:
+        super().__init__(raw, status, code)
+        
+        self.ip: str = raw["ip"]
+        self.type: Literal["ipv4", "ipv6"] = raw["type"]
+        self.continent_code: str = raw["continent_code"]
+        self.continent_name: str = raw["continent_name"]
+        self.country_code: str = raw["country_code"]
+        self.country_name: str = raw["country_name"]
+        self.is_eu: bool = bool(raw["is_eu"])
+        self.geoname_id: int = raw["geoname_id"]
+        self.city: str = raw["city"]
+        self.region: str = raw["region"]
+        self.latitude: float = raw["lat"]
+        self.longitude: float = raw["lon"]
+        self.tz_id: str = raw["tz_id"]
+        self.localtime_epoch: int = raw["localtime_epoch"]
+        self.localtime_formatted: str = raw["localtime"]